+++ contracts/SinglePlus.sol
@@ -128,7 +128,7 @@ contract SinglePlus is ISinglePlus, Plus, ReentrancyGuardUpgradeable {
         // Special handling of -1 is required here in order to fully redeem all shares, since interest
         // will be accrued between the redeem transaction is signed and mined.
         uint256 _share;
-        if (_amount == uint256(-1)) {
+        if (_amount == uint256(int256(-1))) {
             _share = userShare[msg.sender];
             _amount = _share.mul(index).div(WAD);
         } else {
